moved timeline code into TimelineService
