+++ src/Service/TimelineService.php
@@ -4,25 +4,26 @@ require_once "NoteService.php";
 require_once "PersonService.php";
 require_once "EventService.php";
 
+use ChurchCRM\NoteQuery;
+use ChurchCRM\PersonQuery;
+
 class TimelineService
 {
 
   private $personService;
-  private $noteService;
   private $eventService;
 
   public function __construct()
   {
     $this->personService = new PersonService();
-    $this->noteService = new NoteService();
     $this->eventService = new EventService();
   }
 
   function getForFamily($familyID)
   {
     $timeline = array();
-
-    $notes = $this->noteService->getNotesByFamily($familyID, $_SESSION['bAdmin'], $_SESSION['iUserID']);
+    $rawNotes = NoteQuery::create()->findByFamId($familyID);
+    $notes = $this->convertNotes($rawNotes, $_SESSION['bAdmin']);
     foreach ($notes as $note) {
       $item = $this->createTimeLineItem($note["type"], $note["lastUpdateDatetime"],
         "by " . $note["lastUpdateByName"], "", $note["text"],
@@ -45,8 +46,8 @@ class TimelineService
   function getForPerson($personID)
   {
     $timeline = array();
-
-    $notes = $this->noteService->getNotesByPerson($personID, $_SESSION['bAdmin'], $_SESSION['iUserID']);
+    $rawNotes = NoteQuery::create()->findByPerId($personID);
+    $notes = $this->convertNotes($rawNotes, $_SESSION['bAdmin']);
     foreach ($notes as $note) {
       $item = $this->createTimeLineItem($note["type"], $note["lastUpdateDatetime"],
         "by " . $note["lastUpdateByName"], "", $note["text"],
@@ -108,5 +109,39 @@ class TimelineService
 
     return $item;
   }
+
+  function convertNotes($notes, $admin)
+  {
+    $notesArray = array();
+    foreach ($notes as $rawNote) {
+      // if the user is not admin, ensure the note is not private or it is created by current user
+      if ($admin || $rawNote->isVisable($_SESSION['iUserID'])) {
+
+        $note['id'] = $rawNote->getId();
+        $note['private'] = $rawNote->getPrivate();
+        $note['text'] = $rawNote->getText();
+        $note['type'] = $rawNote->getType();
+
+        if ($rawNote->getDateLastEdited() != null) {
+          $note['lastUpdateDatetime'] = $rawNote->getDateLastEdited()->format('Y-m-d H:i:s');
+          $note['lastUpdateById'] = $rawNote->getEditedBy();
+        } else {
+          $note['lastUpdateDatetime'] = $rawNote->getDateEntered()->format('Y-m-d H:i:s');
+          $note['lastUpdateById'] = $rawNote->getEnteredBy();
+        }
+
+        $person = PersonQuery::create()->findPk($note['lastUpdateById']);
+
+        if ($person != null) {
+          $note['lastUpdateByName'] = $person->getFullName();
+        } else {
+          $note['lastUpdateByName'] = "unknown?";
+        }
+        array_push($notesArray, $note);
+      }
+    }
+    return $notesArray;
+  }
+
 
 }
